neighbourhood function with potential to give larger neighbourhood
right now the smaller neighbourhood is switched on again. The larger one gives better results but takes ten times the time the smaller one takes. Still worth it maybe?

Fact is, the amount of invalid neighbour that are checked grows like crazy, if we find a way to optimize that... Would be amazing.